Crate iter_extended

Functions

btree_map

Equivalent to .into_iter().map(f).collect::<BTreeMap<K, V>>()

partition_results

Given an iterator over a Result, filter out the Ok values from the Err values and return both in separate Vecs. Unlike other collect-like functions over Results, this function will always consume the entire iterator.

try_btree_map

Equivalent to .into_iter().map(f).collect::<Result<BTreeMap<_, >,>>()

try_vecmap

Equivalent to .into_iter().map(f).collect::<Result<Vec<>,>>()

vecmap

Equivalent to .into_iter().map(f).collect::<Vec<_>>()